# public/swagger/users.json
{
    "openapi": "3.0.0",
    "servers": [
        {
            "description": "Production Backend",
            "url": "http://localhost:5000/v0/users"
        },
        {
            "description": "Local Development Backend",
            "url": "http://localhost:5000"
        }
    ],
    "info": {
        "version": "0.0.1-oas3",
        "title": "Make My Match",
        "description": "# Handle the user\n CRUD APIS of Make My Match."
    },
    "security": [
        {
            "application": [
                "read",
                "write"
            ]
        },
        {
            "ApiKeyAuth": []
        }
    ],
    "paths": {
...
        "/login": {
            "summary": "Get jwt token with your username and password",
            "description": "Send me your credentials, I'll tell you if you can enter",
            "post": {
                "operationId": "postAuthLogin",
                "summary": "Validate your user credentials to enter CookinAt' app",
                "description": "Given email and password, create and return jwt token and DB id",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/Login"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Ok"
                    },
                    "401": {
                        "description": "Email or password is wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnAuthorized"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequest"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "tags": [
                    "App Authorization"
                ]
            }
        },
        "/social/login": {
            "summary": "Create user in database of Users, if user is already there it will provide the token to you.",
            "description": "If you want access with social media authorization, come here first",
            "post": {
                "operationId": "postAuthSocialMedia",
                "summary": "Get access to 'Make My Match' App creating a User",
                "description": "In the request body you give social_media_id and social_media_name.\nWe store that data and\nsent you a token code.\nThen you can get enter in app.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/SocialLogin"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/SocialLogin200"
                    },
                    "422": {
                        "description": "Errors in Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError",
                        "description": "Internal Server Error"
                    }
                },
                "tags": [
                    "App Authorization"
                ]
            }
        },
...
        "/signup": {
            "summary": "Create user in database of Users",
            "description": "If you want access come here first",
            "post": {
                "operationId": "postAuthSignup",
                "summary": "Get access to 'Make My Match' App creating a User",
                "description": "In the App you give username, email, mobile_number and password.\nWe store that data, encrypt your password and\nsent you a token code.\nThen you can get enter in app.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/Signup"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/SignUp200"
                    },
                    "409": {
                        "$ref": "#/components/responses/Email409"
                    },
                    "422": {
                        "description": "Errors in Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError",
                        "description": "Internal Server Error"
                    }
                },
                "tags": [
                    "App Authorization"
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "authorization"
            }
        },
        "schemas": {
            "Login": {
                "type": "object",
                "required": [
                    "username",
                    "password"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "Complete username without spaces",
                        "example": "test"
                    },
                    "password": {
                        "type": "string",
                        "description": "password with more than 6",
                        "example": "password123"
                    }
                }
            },
            "Signup": {
                "type": "object",
                "required": [
                    "username",
                    "email",
                    "password",
                    "mobile_number"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "Unique user name to identify it in app",
                        "example": "johan"
                    },
                    "password": {
                        "type": "string",
                        "description": "user password",
                        "example": "doe"
                    },
                    "email": {
                        "type": "string",
                        "description": "email without spaces",
                        "example": "example@mail.com"
                    },
                    "mobile_number": {
                        "type": "number",
                        "description": "mobile number without spaces",
                        "example": "1234567890"
                    }
                }
            },
            "SocialLogin": {
                "type": "object",
                "required": [
                    "social_media_name",
                    "social_media_id"
                ],
                "properties": {
                    "social_media_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "social_media_id": {
                        "type": "string",
                        "example": "1234567890"
                    }
                }
            },
...
            "Ok": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "number",
                        "description": "Response status"
                    },
                    "code": {
                        "type": "number",
                        "description": "Response body code"
                    },
                    "data": {
                        "type": "object",
                        "description": "object with valuable data"
                    },
                    "message": {
                        "type": "string",
                        "description": "Explanatory message"
                    },
                    "emptyKeys": {
                        "type": "array"
                    },
                    "error": {
                        "type": "boolean"
                    }
                },
                "example": {
                    "status": 200,
                    "code": 1,
                    "data": {
                        "_id": "5d4537d91bf2f20cbe51eb70",
                        "username": "test2",
                        "mobile_number": "1234567890",
                        "deletedStatus": 0,
                        "status": 0,
                        "location": {
                            "type": "Point",
                            "coordinates": [
                                "30.7333",
                                "76.7794"
                            ]
                        },
                        "user_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI1ZDQ1MzdkOTFiZjJmMjBjYmU1MWViNzAiLCJ1c2VybmFtZSI6InRlc3QyIiwibW9iaWxlX251bWJlciI6IjEyMzQ1Njc4OTAiLCJkZWxldGVkU3RhdHVzIjowLCJzdGF0dXMiOjAsImxvY2F0aW9uIjp7InR5cGUiOiJQb2ludCIsImNvb3JkaW5hdGVzIjpbIjMwLjczMzMiLCI3Ni43Nzk0Il19LCJ1c2VyX3Rva2VuIjoiZXlKaGJHY2lPaUpJVXpJMU5pSXNJblI1Y0NJNklrcFhWQ0o5LmV5SjFjMlZ5Ym1GdFpTSTZJblJsYzNReUlpd2liVzlpYVd4bFgyNTFiV0psY2lJNklqRXlNelExTmpjNE9UQWlMQ0prWld4bGRHVmtVM1JoZEhWeklqb3dMQ0p6ZEdGMGRYTWlPakFzSW14dlkyRjBhVzl1SWpwN0luUjVjR1VpT2lKUWIybHVkQ0lzSW1OdmIzSmthVzVoZEdWeklqcGJiblZzYkN4dWRXeHNYWDBzSW1saGRDSTZNVFUyTkRneE56TTJPSDAuRXc5VG1LbDBUWjVES3RoMEhmV1ptV0VNRmVpakQyLUVKZTVXajd3a09tQSIsImFib3V0X21lIjoiTG9yZW0gSXBzdW0gaXMgc2ltcGx5IGR1bW15IHRleHQgb2YgdGhlIHByaW50aW5nIGFuZCB0eXBlc2V0dGluZyBpbmR1c3RyeS4gTG9yZW0gSXBzdW0gaGFzIGJlZW4gdGhlIGluZHVzdHJ5J3Mgc3RhbmRhcmQgZHVtbXkgdGV4dCBldmVyIHNpbmNlIHRoZSAxNTAwcywgd2hlbiBhbiB1bmtub3duIHByaW50ZXIgdG9vayBhIGdhbGxleSBvZiB0eXBlIGFuZCBzY3JhbWJsZWQgaXQgdG8gbWFrZSBhIHR5cGUgc3BlY2ltZW4gYm9vay4iLCJleWVfY29sb3IiOiJibGFjayIsImhhaXJfY29sb3IiOiJibGFjayIsImlzUHJpdmF0ZSI6IkZBTFNFIiwib2NjdXBhdGlvbiI6IkRldmVsb3BlciIsInByb2ZpbGVfaW1hZ2VfdXJsIjoicHJvZmlsZV9pbWFnZV91cmwiLCJyZWxhdGlvbnNoaXBfc3RhdHVzIjoiU2luZ2xlIiwid2hhdF9oZV9pc19sb29raW5nX2ZvciI6IkxvcmVtIElwc3VtIGlzIHNpbXBseSBkdW1teSB0ZXh0IG9mIHRoZSBwcmludGluZyBhbmQgdHlwZXNldHRpbmcgaW5kdXN0cnkuIiwiaWF0IjoxNTY1NjkzMjQ0fQ.vO-nrd9IkZb5d1y_c_0mDQc1WoNQDETI1vM_NGwz_Uk",
                        "about_me": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book.",
                        "eye_color": "black",
                        "hair_color": "black",
                        "isPrivate": "FALSE",
                        "occupation": "Developer",
                        "profile_image_url": "profile_image_url",
                        "relationship_status": "Single",
                        "what_he_is_looking_for": "Lorem Ipsum is simply dummy text of the printing and typesetting industry."
                    },
                    "message": "Success",
                    "emptyKeys": null,
                    "error": false
                }
            },
            "ServerError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "number"
                    },
                    "status": {
                        "type": "number"
                    },
                    "data": {
                        "type": "array"
                    },
                    "message": {
                        "type": "string"
                    },
                    "emptyKeys": {
                        "type": "array"
                    },
                    "error": {
                        "type": "boolean"
                    }
                },
                "example": {}
            },
            "ClientError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "number"
                    },
                    "status": {
                        "type": "number"
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    },
                    "emptyKeys": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "error": {
                        "type": "boolean"
                    }
                },
                "example": {
                    "code": 1,
                    "status": 409,
                    "data": {},
                    "message": "Present",
                    "emptyKeys": null,
                    "error": true
                }
            },
...
            "User201": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "number"
                    },
                    "code": {
                        "type": "number"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "_id": {
                                "type": "string",
                                "description": "Unique id"
                            },
                            "username": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "mobile_number": {
                                "type": "number"
                            },
                            "deletedStatus": {
                                "type": "number"
                            },
                            "status": {
                                "type": "number"
                            },
                            "user_token": {
                                "type": "number",
                                "description": "Token to access the user data."
                            }
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "emptyKeys": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "error": {
                        "type": "boolean"
                    }
                }
            },
            "SocialLoginUser201": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "number"
                    },
                    "code": {
                        "type": "number"
                    },
                    "data": {
                        "_id": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        },
                        "loginType": {
                            "type": "string"
                        },
                        "deletedStatus": {
                            "type": "number"
                        },
                        "status": {
                            "type": "number"
                        },
                        "user_token": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "emptyKeys": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "error": {
                        "type": "boolean"
                    }
                }
            },
...
            "UnAuthorized": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "number"
                    },
                    "status": {
                        "type": "number"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "message": {
                                "type": "string"
                            }
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "emptyKeys": {
                        "type": "array"
                    },
                    "error": {
                        "type": "boolean"
                    }
                },
                "example": {
                    "code": 1,
                    "status": 401,
                    "data": {
                        "message": "Either username or password is incorrect."
                    },
                    "message": "Not Authorized",
                    "emptyKeys": null,
                    "error": true
                }
            },
            "ValidationError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "number"
                    },
                    "status": {
                        "type": "number"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "emptyKeys": {
                        "type": "object",
                        "properties": {
                            "error": {
                                "type": "string"
                            },
                            "message": {
                                "type": "array",
                                "items": {
                                    "fieldName": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "error": {
                        "type": "boolean"
                    }
                },
                "example": {
                    "code": 1,
                    "status": 422,
                    "data": [],
                    "message": "ValidationError",
                    "emptyKeys": {
                        "error": "type_mismatched",
                        "message": [
                            {
                                "fieldName": "username",
                                "type": "string"
                            }
                        ]
                    },
                    "error": true
                }
            },
            "BadRequest": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "number"
                    },
                    "status": {
                        "type": "number"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "message": {
                                "type": "string"
                            }
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "emptyKeys": {
                        "type": "array"
                    },
                    "error": {
                        "type": "boolean"
                    }
                },
                "example": {
                    "code": 1,
                    "status": 400,
                    "data": {
                        "message": "Authorization token is not found."
                    },
                    "message": "Bad REQUEST",
                    "emptyKeys": null,
                    "error": true
                }
            }
        },
        "requestBodies": {
...
            "Login": {
                "description": "User credentials needed",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/Login"
                        }
                    },
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Login"
                        }
                    }
                }
            },
            "Signup": {
                "description": "User related data needed",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/Signup"
                        }
                    },
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Signup"
                        }
                    }
                }
            },
            "SocialLogin": {
                "description": "User related data needed",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/SocialLogin"
                        }
                    },
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SocialLogin"
                        }
                    }
                }
            },
...
        },
        "responses": {
            "ServerError": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ServerError"
                        }
                    }
                }
            },
...
            "Email409": {
                "description": "Email already exists",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ClientError"
                        }
                    }
                }
            },
            "Ok": {
                "description": "The petition returned successfuly",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Ok"
                        }
                    }
                }
            },
...
            "SignUp200": {
                "description": "User Sign Up successfully (email and username inside of data) and also you can consider youself as logged in user. #In short you got your token.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/User201"
                        }
                    }
                },
                "links": {
                    "getEmail": {
                        "operationId": "postAuthLogin",
                        "parameters": {
                            "userId": "$response.body#/data/email"
                        },
                        "description": "The `email` value returned in the response can be used as the `email` parameter in `POST /auth/login`.\n"
                    }
                }
            },
            "SocialLogin200": {
                "description": "User Logged in Successfuly",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SocialLoginUser201"
                        }
                    }
                },
                "links": {
                    "getEmail": {
                        "operationId": "postAuthLogin",
                        "parameters": {
                            "userId": "$response.body#/data/email"
                        },
                        "description": "The `email` value returned in the response can be used as the `email` parameter in `POST /auth/login`.\n"
                    }
                }
            },
...
        },
...
    }
}